Refactor AI helper functions
